+++ R/bcdc-web-services.R
@@ -11,12 +11,17 @@
 # See the License for the specific language governing permissions and limitations under the License.
 
 
-#' Query data from the B.C. Web Service
+#' Query data from the B.C. Web Feature Service
 #'
-#' Queries features from the B.C. Web Service. The data must be available as
-#' a Web Service. See `bcdc_get_record(record)$resources`). If the record is greater than 10000 rows,
-#' the response will be paginated. If you are querying layers of this size, expect
-#' that the request will take quite a while.
+#' Queries features from the B.C. Web Feature Service. See
+#' [bcdc_tidy_resources()] - if a resource has a value of
+#' `"wms"` in the `format` column it is available as a Web
+#' Feature Service, and you can query and download it
+#' using `bcdc_query_geodata()`. The response will be
+#' paginated if the number of features is above the number
+#' set by the `bcdata.single_download_limit` option.
+#' Please see [bcdc_options()] for defaults and more
+#' information.
 #'
 #' Note that this function doesn't actually return the data, but rather an
 #' object of class `bcdc_promise`, which includes all of the information
@@ -40,31 +45,45 @@
 #'
 #' \donttest{
 #' # Returns a bcdc_promise, which can be further refined using filter/select:
-#' bcdc_query_geodata("bc-airports", crs = 3857)
+#' try(
+#'   bcdc_query_geodata("bc-airports", crs = 3857)
+#' )
 #'
 #' # To obtain the actual data as an sf object, collect() must be called:
-#' bcdc_query_geodata("bc-airports", crs = 3857) %>%
-#'   filter(PHYSICAL_ADDRESS == 'Victoria, BC') %>%
-#'   collect()
+#' try(
+#'   bcdc_query_geodata("bc-airports", crs = 3857) %>%
+#'     filter(PHYSICAL_ADDRESS == 'Victoria, BC') %>%
+#'     collect()
+#' )
 #'
-#' bcdc_query_geodata("ground-water-wells") %>%
-#'   filter(OBSERVATION_WELL_NUMBER == 108) %>%
-#'   select(WELL_TAG_NUMBER, INTENDED_WATER_USE) %>%
-#'   collect()
+#' try(
+#'   bcdc_query_geodata("groundwater-wells") %>%
+#'     filter(OBSERVATION_WELL_NUMBER == "108") %>%
+#'     select(WELL_TAG_NUMBER, INTENDED_WATER_USE) %>%
+#'     collect()
+#' )
 #'
 #' ## A moderately large layer
-#' bcdc_query_geodata("bc-environmental-monitoring-locations")
-#' bcdc_query_geodata("bc-environmental-monitoring-locations") %>%
-#'   filter(PERMIT_RELATIONSHIP == "DISCHARGE")
+#' try(
+#'   bcdc_query_geodata("bc-environmental-monitoring-locations")
+#' )
+#'
+#' try(
+#'   bcdc_query_geodata("bc-environmental-monitoring-locations") %>%
+#'     filter(PERMIT_RELATIONSHIP == "DISCHARGE")
+#' )
 #'
 #'
 #' ## A very large layer
-#' bcdc_query_geodata("terrestrial-protected-areas-representation-by-biogeoclimatic-unit")
+#' try(
+#'   bcdc_query_geodata("terrestrial-protected-areas-representation-by-biogeoclimatic-unit")
+#' )
 #'
 #' ## Using a BCGW name
-#' bcdc_query_geodata("WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW")
+#' try(
+#'   bcdc_query_geodata("WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW")
+#' )
 #' }
-#'
 #' @export
 bcdc_query_geodata <- function(record, crs = 3005) {
   if (!has_internet()) stop("No access to internet", call. = FALSE) # nocov
@@ -118,7 +137,7 @@ bcdc_query_geodata.character <- function(record, crs = 3005) {
 #' @export
 bcdc_query_geodata.bcdc_record <- function(record, crs = 3005) {
   if (!any(wfs_available(record$resource_df))) {
-    stop("No Web Service resource available for this data set.",
+    stop("No Web Feature Service resource available for this data set.",
          call. = FALSE
     )
   }
@@ -142,18 +161,29 @@ bcdc_query_geodata.bcdc_record <- function(record, crs = 3005) {
                        cols_df = cols_df))
 }
 
-#' Get map from the B.C. Web Service
+#' Get preview map from the B.C. Web Map Service
+#'
+#' Note this does not return the actual map features, rather
+#' opens an image preview of the layer in a
+#' [Leaflet](https://rstudio.github.io/leaflet/) map window
 #'
 #'
 #' @inheritParams bcdc_get_data
 #'
 #' @examples
 #' \donttest{
-#' bcdc_preview("regional-districts-legally-defined-administrative-areas-of-bc")
-#' bcdc_preview("points-of-well-diversion-applications")
+#' try(
+#'   bcdc_preview("regional-districts-legally-defined-administrative-areas-of-bc")
+#' )
+#'
+#' try(
+#'   bcdc_preview("points-of-well-diversion-applications")
+#' )
 #'
 #' # Using BCGW name
-#' bcdc_preview("WHSE_LEGAL_ADMIN_BOUNDARIES.ABMS_REGIONAL_DISTRICTS_SP")
+#' try(
+#'   bcdc_preview("WHSE_LEGAL_ADMIN_BOUNDARIES.ABMS_REGIONAL_DISTRICTS_SP")
+#' )
 #' }
 #' @export
 bcdc_preview <- function(record) { # nocov start