+++ R/describe-feature.R
@@ -10,26 +10,32 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
 
-#' Describe the attributes of a Web Service feature
+#' Describe the attributes of a Web Feature Service
 #'
-#' Describe the attributes of column of a record accessed through the Web Service.
+#' Describe the attributes of column of a record accessed through the Web Feature Service.
 #' This can be a useful tool to examine a layer before issuing a query with `bcdc_query_geodata`.
 #'
 #' @return
-#' `bcdc_describe_features` returns a tibble describing the attributes of a B.C. Data Catalogue record.
+#' `bcdc_describe_feature` returns a tibble describing the attributes of a B.C. Data Catalogue record.
 #' The tibble returns the following columns:
 #'   - col_name: attributes of the feature
-#'   - sticky: whether a column can be separated from the record in a Web Service call via the `dplyr::select` method
+#'   - sticky: whether a column can be separated from the record in a Web Feature Service call via the `dplyr::select` method
 #'   - remote_col_type: class of what is return by the web feature service
 #'   - local_col_type: the column class in R
+#'   - column_comments: additional metadata specific to that column
 #'
 #' @inheritParams bcdc_query_geodata
 #' @export
 #'
 #' @examples
 #' \donttest{
-#'  bcdc_describe_feature("bc-airports")
-#'  bcdc_describe_feature("WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW")
+#' try(
+#'   bcdc_describe_feature("bc-airports")
+#' )
+#'
+#' try(
+#'   bcdc_describe_feature("WHSE_IMAGERY_AND_BASE_MAPS.GSR_AIRPORTS_SVW")
+#' )
 #' }
 #'
 #' @export
@@ -48,7 +54,9 @@ bcdc_describe_feature.default <- function(record) {
 bcdc_describe_feature.character <- function(record){
 
   if (is_whse_object_name(record)) {
-    return(feature_helper(record))
+    bgc <- bcdc_get_wfs_records()
+    cat_record <- bcdc_get_record(bgc$cat_url[grepl(record, bgc$whse_name)])
+    return(obj_desc_join(record, cat_record$details))
   }
 
   bcdc_describe_feature(bcdc_get_record(record))
@@ -59,12 +67,12 @@ bcdc_describe_feature.character <- function(record){
 bcdc_describe_feature.bcdc_record <- function(record){
 
   if (!any(wfs_available(record$resource_df))) {
-    stop("No WMS/WFS resource available for this data set.",
+    stop("No WFS resource available for this data set.",
          call. = FALSE
     )
   }
+  obj_desc_join(record$layer_name, record$details)
 
-  feature_helper(record$layer_name)
 }
 
 parse_raw_feature_tbl <- function(query_list){
@@ -105,15 +113,11 @@ feature_helper <- function(whse_name){
   xml_df <- parse_raw_feature_tbl(query_list)
   geom_type <- attr(xml_df, "geom_type")
 
-  ## Identify geometry column and move to last
-  # xml_df[xml_df$type == geom_type, "name"] <- "geometry"
-  # xml_df <- dplyr::bind_rows(xml_df[xml_df$name != "geometry",],
-  #                            xml_df[xml_df$name == "geometry",])
 
   ## Fix logicals
   xml_df$nillable = ifelse(xml_df$nillable == "true", TRUE, FALSE)
-
   xml_df <- xml_df[, c("name", "nillable", "type")]
+
   ## Add the id_row back into the front
   xml_df <- dplyr::bind_rows(id_row, xml_df)
   colnames(xml_df) <- c("col_name", "sticky", "remote_col_type")
@@ -124,4 +128,10 @@ feature_helper <- function(whse_name){
 
 
 
-
+obj_desc_join <- function(x, y) {
+  dplyr::left_join(
+    feature_helper(x),
+    y[,c("column_comments", "column_name")],
+    by = c("col_name" = "column_name")
+  )
+}