+++ tests/testthat/test-options.R
@@ -19,10 +19,29 @@ test_that("bcdc_options() returns a tibble",{
 
 
 test_that("bcdata.chunk_limit",{
-  original_value <- options(bcdata.chunk_limit = 10000)
+  withr::local_options(list(bcdata.chunk_limit = 10000))
   expect_error(check_chunk_limit())
+})
 
-  ## cleaning up. not sure if this is required but likely not a bad practice.
-  options(bcdata.chunk_limit = original_value$bcdata.chunk_limit)
+test_that("bcdata.single_download_limit", {
+  skip_on_cran()
+  withr::local_options(list(bcdata.single_download_limit = 1))
+  expect_message(
+    bcdc_get_data(record = '76b1b7a3-2112-4444-857a-afccf7b20da8', resource =
+                    '4d0377d9-e8a1-429b-824f-0ce8f363512c'),
+    "paginated"
+  )
 
+})
+
+test_that("bcdata.single_download_limit can be changed",{
+  withr::local_options(list(bcdata.single_download_limit = 13))
+  expect_equal(getOption("bcdata.single_download_limit"), 13)
+})
+
+test_that("bcdc_single_download_limit returns a number",{
+  skip_on_cran()
+  skip_if_net_down()
+  lt <- bcdc_single_download_limit()
+  expect_type(lt, "integer")
 })