+++ NEWS.md
@@ -1,8 +1,33 @@
-# bcdata 0.2.0.9000
+# bcdata 0.2.2.9000
+### IMPROVEMENTS
+- Setting the `bcdata.single_download_limit` limit dynamically from the getCapabilities endpoint. #256
+- `bcdc_describe_feature` now joins an object description column to the returned object to provide more information about a field directly in R. #241, #259
+- Better documentation and information surrounding the `bcdata.max_geom_pred_size` option. #243, #258 
+- Add new function `bcdc_check_geom_size` so users can check prior to submitting a WFS request with `filter` to see 
+if the request will require a bounding box conversion. #243, #258
+- Better documentation and messaging about when and why paginated requests are required by `bcdc_query_geodata()`. #240, #264
+- Better documentation and print method for what records are suitable for use with `bcdc_query_geodata()`. #265, #267
+
+# bcdata 0.2.2
+### IMPROVEMENTS
+* Added `bcdc_list_groups` and `bcdc_list_group_records` to provide the ability to query on the group endpoint of the catalogue API. #234
+* Added new option `bcdata.single_download_limit` to enable setting the maximum number of records an object can be before forcing a paginated download (#252)
+
+### BUG FIXES
+* Fixed bug in `collect.bcdc_promise` where the wrong parameter name in `crul::Paginator$new()` resulted in an error in paginated wfs requests (#250, thanks @meztez)
+* Fixed a bug where the name of `bcdata.chunk_limit` option had a typo, so that it was not actually used properly (#252)
+
+# bcdata 0.2.1
 
 ### BUG FIXES
 * Remove link for pipe documentation for simplicity.
 * Fixed bug where using many `as.` functions (e.g., `as.Date()`, `as.character()`, `as.numeric()`) in a filter statement would fail. (#218, #219)
+
+### MAINTENANCE
+* Updated internal SQL translation to use `DBI` S4 generics (`DBI::dbQuoteIdentifier()` is now used instead of 
+  `dbplyr::sql_escape_ident()` and `DBI::dbQuoteString()` instead of `dbplyr::sql_escape_string()`), to comply 
+  with upcoming `dbplyr` 2.0 release (#225, #225; https://github.com/tidyverse/dbplyr/issues/385)
+* Wrapped all examples that call web resources in `try()` to avoid spurious check failures (#229).
 
 # bcdata 0.2.0
 
@@ -17,7 +42,7 @@
 * When reading in excel files, `bcdc_get_data` now outputs a messages indicating the presence and names of any sheets (#190)
 * `bcdc_get_data()` & `bcdc_query_geodata()` will now work with full B.C. data catalogue url including resource (#125, #196)
 * `bcdc_sf` objects now have an `time_downloaded` attribute
-* Authorized B.C. Data Catalogue editors can now authenticate with the catalague by setting 
+* Authorized B.C. Data Catalogue editors can now authenticate with the catalogue by setting 
 a `BCDC_KEY` environment variable with their catalogue API token (https://github.com/bcgov/bcdata#bcdc-authentication; #208).
 
 ### BUG FIXES
@@ -42,7 +67,7 @@ a `BCDC_KEY` environment variable with their catalogue API token (https://github
 * Now precompiling vignettes so that queries are submitted locally and no actually requests are made from CRAN (#151)
 * Fix `NOTE: Namespace in Imports field not imported from: ‘methods’` error on CRAN (#145)
 * Fixed a bug where functions nested inside geometry predicates were not evaluated (#146, #154)
-* Fixed a bug where `DWITHIN` wasn't working because `units` needed to be unqoted (#154)
+* Fixed a bug where `DWITHIN` wasn't working because `units` needed to be unquoted (#154)
 * Fixed a bug where `BBOX()` used in a `filter()` statement combined with `bcdc_query_geodata()` did not work (#135, #137, #131)
 * Fixed a bug where layer names with a number in them would not work in `bcdc_query_geodata()` (#126, #127)
 
@@ -57,7 +82,7 @@ a `BCDC_KEY` environment variable with their catalogue API token (https://github
 * Better and more informative error message when experiencing http failures occur (#121)
 * Added print methods for `show_query`
 * Change examples to donttest
-* Added verbose argument to `bcdc_get_record` to enable supressing console writing
+* Added verbose argument to `bcdc_get_record` to enable suppressing console writing
 
 # bcdata 0.1.0